Moved imagequery to imagepage
