+++ src/schema.graphql
@@ -440,17 +440,6 @@ type ImageAltText {
   language: String!
 }
 
-type ImageCopyRight {
-  agreementId: Int
-  creators: [Contributor!]!
-  license: License!
-  origin: String!
-  processors: [Contributor!]!
-  rightsholders: [Contributor!]!
-  validFrom: String
-  validTo: String
-}
-
 type ImageDimensions {
   height: Int!
   width: Int!
@@ -497,16 +486,16 @@ type ImageMetaInformation {
   title: String!
 }
 
-type ImageMetaInformationV3 {
+type ImageMetaInformationV2 {
   alttext: ImageAltText!
   caption: Caption!
   contentType: String!
-  copyright: ImageCopyRight!
+  copyright: Copyright!
   created: String!
   createdBy: String!
   editorNotes: [EditorNote!]
   id: String!
-  imageDimensions: ImageDimensions!
+  imageDimensions: ImageDimensions
   imageUrl: String!
   metaUrl: String!
   modelRelease: String!
@@ -809,7 +798,7 @@ type Query {
     resourceTypes: String
     subjects: String
   ): [GroupSearch!]
-  image(id: String!): ImageMetaInformationV3!
+  image(id: String!): ImageMetaInformationV2
   learningpath(pathId: String!): Learningpath
   listingPage(subjects: String): ListingPage
   personalData: MyNdlaPersonalData!
