+++ src/util/copyrightHelpers.ts
@@ -1,3 +1,11 @@
+/**
+ * Copyright (c) 2023-present, NDLA.
+ *
+ * This source code is licensed under the GPLv3 license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
 import { getGroupedContributorDescriptionList } from '@ndla/licenses';
 import { GQLContributor } from '../graphqlTypes';
 
@@ -6,10 +14,10 @@ export const getPrioritizedAuthors = (authors: {
   rightsholders: GQLContributor[];
   processors: GQLContributor[];
 }): GQLContributor[] => {
-  const { creators, rightsholders, processors } = authors;
+  const { creators = [], rightsholders = [], processors = [] } = authors;
 
   if (creators.length || rightsholders.length) {
-    return [...creators, ...rightsholders];
+    return creators.concat(rightsholders);
   }
   return processors;
 };
